# prototype/notification-manager/docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Notification Manager API
  description: |
    RESTful API for querying notification data from the L'Aquila Digital Twin system.
    Notifications are consumed from Kafka and stored in MongoDB, made available through 
    this API with advanced filtering, sorting, and pagination capabilities.
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com

servers:
  - url: http://localhost:3001/api
    description: Local development server
  - url: https://notifications.example.com/api
    description: Production server

tags:
  - name: Health
    description: System health check endpoints
  - name: Notifications
    description: Notification query endpoints

paths:
  /health:
    get:
      tags:
        - Health
      summary: Health check
      description: Check if the API server and MongoDB are operational
      operationId: getHealth
      responses:
        '200':
          description: System is healthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
        '503':
          description: System is unhealthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnhealthyResponse'

  /notifications:
    get:
      tags:
        - Notifications
      summary: Get notifications
      description: Retrieve notifications with optional filtering, sorting, and pagination
      operationId: getNotifications
      parameters:
        - name: severity
          in: query
          description: Filter by severity level(s). Multiple values can be comma-separated.
          required: false
          schema:
            oneOf:
              - type: string
                enum: [info, warning, error, critical]
              - type: string
                pattern: '^(info|warning|error|critical)(,(info|warning|error|critical))*$'
          examples:
            single:
              value: warning
              summary: Single severity
            multiple:
              value: warning,critical
              summary: Multiple severities
        - name: source
          in: query
          description: Filter by source(s). Multiple values can be comma-separated.
          required: false
          schema:
            type: string
          examples:
            single:
              value: sensor-network
            multiple:
              value: sensor-network,traffic-monitoring
        - name: timestampFrom
          in: query
          description: Start of time range (inclusive)
          required: false
          schema:
            type: string
            format: date-time
          example: '2024-12-01T00:00:00Z'
        - name: timestampTo
          in: query
          description: End of time range (inclusive)
          required: false
          schema:
            type: string
            format: date-time
          example: '2024-12-08T23:59:59Z'
        - name: limit
          in: query
          description: Maximum number of results to return
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 50
        - name: offset
          in: query
          description: Number of results to skip for pagination
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: sortBy
          in: query
          description: Field to sort results by
          required: false
          schema:
            type: string
            enum: [timestamp, createdAt, severity]
            default: timestamp
        - name: sortOrder
          in: query
          description: Sort order (ascending or descending)
          required: false
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: Notifications retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationListResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /notifications/{id}:
    get:
      tags:
        - Notifications
      summary: Get notification by ID
      description: Retrieve a specific notification by its MongoDB ObjectId
      operationId: getNotificationById
      parameters:
        - name: id
          in: path
          required: true
          description: MongoDB ObjectId (24-character hexadecimal string)
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
          example: 507f1f77bcf86cd799439011
      responses:
        '200':
          description: Notification found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationResponse'
        '400':
          description: Invalid ID format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Invalid notification ID format
        '404':
          description: Notification not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Notification not found
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    HealthResponse:
      type: object
      properties:
        status:
          type: string
          enum: [healthy]
          example: healthy
        timestamp:
          type: string
          format: date-time
          example: '2024-12-08T10:30:00.000Z'
        mongodb:
          type: string
          enum: [connected]
          example: connected

    UnhealthyResponse:
      type: object
      properties:
        status:
          type: string
          enum: [unhealthy]
          example: unhealthy
        timestamp:
          type: string
          format: date-time
          example: '2024-12-08T10:30:00.000Z'
        error:
          type: string
          example: Connection timeout

    Notification:
      type: object
      required:
        - _id
        - message
        - severity
        - source
        - timestamp
      properties:
        _id:
          type: string
          description: MongoDB ObjectId
          example: 507f1f77bcf86cd799439011
        message:
          type: string
          description: Notification message content
          maxLength: 5000
          example: High traffic congestion detected on Via XX Settembre
        severity:
          type: string
          description: Notification severity level
          enum: [info, warning, error, critical]
          example: warning
        source:
          type: string
          description: Source system or component that generated the notification
          example: traffic-monitoring
        timestamp:
          type: string
          format: date-time
          description: When the notification was generated (ISO 8601)
          example: '2024-12-08T10:15:30.500Z'
        createdAt:
          type: string
          format: date-time
          description: When the record was created in the database
          example: '2024-12-08T10:15:31.000Z'
        updatedAt:
          type: string
          format: date-time
          description: When the record was last updated
          example: '2024-12-08T10:15:31.000Z'

    NotificationListResponse:
      type: object
      required:
        - notifications
        - total
        - limit
        - offset
      properties:
        notifications:
          type: array
          items:
            $ref: '#/components/schemas/Notification'
        total:
          type: integer
          description: Total number of notifications matching the filters
          example: 150
        limit:
          type: integer
          description: Maximum number of results returned
          example: 50
        offset:
          type: integer
          description: Number of results skipped
          example: 0

    NotificationResponse:
      type: object
      required:
        - notification
      properties:
        notification:
          $ref: '#/components/schemas/Notification'

    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message
          example: Failed to retrieve notifications
        details:
          type: string
          description: Optional detailed error information
          example: Database connection timeout

  examples:
    InfoNotification:
      value:
        _id: 507f1f77bcf86cd799439011
        message: System maintenance scheduled for 2024-12-10 at 02:00 UTC
        severity: info
        source: system-admin
        timestamp: '2024-12-08T10:00:00.000Z'
        createdAt: '2024-12-08T10:00:01.000Z'
        updatedAt: '2024-12-08T10:00:01.000Z'

    WarningNotification:
      value:
        _id: 507f1f77bcf86cd799439012
        message: High traffic congestion detected on Via XX Settembre
        severity: warning
        source: traffic-monitoring
        timestamp: '2024-12-08T10:15:30.500Z'
        createdAt: '2024-12-08T10:15:31.000Z'
        updatedAt: '2024-12-08T10:15:31.000Z'

    ErrorNotification:
      value:
        _id: 507f1f77bcf86cd799439013
        message: Sensor S-123 failed to report data for 15 minutes
        severity: error
        source: sensor-network
        timestamp: '2024-12-08T10:20:00.000Z'
        createdAt: '2024-12-08T10:20:01.000Z'
        updatedAt: '2024-12-08T10:20:01.000Z'

    CriticalNotification:
      value:
        _id: 507f1f77bcf86cd799439014
        message: Seismic activity detected - magnitude 4.2
        severity: critical
        source: seismic-monitoring
        timestamp: '2024-12-08T10:25:45.123Z'
        createdAt: '2024-12-08T10:25:46.000Z'
        updatedAt: '2024-12-08T10:25:46.000Z'